porcod

+++ common/src/lib.rs
@@ -11,7 +11,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 /// Global IDs for game objects
-pub type ID = u32;
+pub type ID = usize;
 
 /// Represents messages sent from the Server to the Client (S2C).
 #[derive(Serialize, Deserialize, Debug)]
@@ -98,3 +98,4 @@ pub enum StructureTypeE {
 pub struct UnitGroupE {
     pub owner: String,
 }
+
